+++ HISTORY.md
@@ -2,6 +2,8 @@ Release History
 ===============
 
 _MEET_TRUE
+values in check
+check_not/check_not_*
 
 0.0.2 (2023-12-04)
 -------------------